The Rework, pt. 2
Reworked the Ember Staff recipe to feel more like an MFR recipe. Added MFR recipes for creating caminite blend on the carpenters bench, and for frying caminite in the oven.

Applied the same methodology that MFR uses to "weaken" base vanilla gear to all Embers gear. It is config sensitive and will only turn on when MFR's "Weaken Basic Items" config is on.