The Rework, pt. 4
Began modularizing how metals are added in order to respect embers config options to disable certain metals.

IOInit has been reworked so that aluminum, electrum, and nickel will be left out of the metal_types file if their respective config option is turned off. Now, to get it to register this when Minecraft boots the first time...

Entirely rewrite StampInit. Made functions to handle the addition and removal of recipes, including the plate recipes.