The Rework pt. 7
Added a "hcCaminiteFiring" config option as an independent control switch as to whether or not the caminite recipes must be cooked in the oven, rather than the furnace.

Added function to remove the furnace recipes for cooking caminite molds that triggers if Minefantasy disables ceramics in the furnace or if Emberforged enables hcCaminiteFiring

Disabled IOInit to test registry integration.

Tweaked material values a bit

Began implementing a _factories based system for controlling which recipes are active. Started with an AluminumToggle factory class and json file. Looks promising, needs further testing.

Added limestone to flux recipe to the stamper.

Removed a couple json recipes that will need to be reworked in order to respect config options.

Added a firing recipe for the flat stamp, because I forgot until now.