The Rework, pt. 17 the one with the polish
Added MFR style recipe for the Ember Bore, and linked it to its research page.

Added MFR style recipe for the Reaction Chamber, and linked it to its research page.

Added a check so that Ember's copper ingot is only given the research oredict tag when the ingot actually exists

Changed all recipes that use blocks of copper to use the oredict tag blockCopper instead.

Fixed the broken Item Request recipe